+++ package.json
@@ -56,7 +56,7 @@
   "devDependencies": {
     "@babel/plugin-transform-modules-commonjs": "^7.18.2",
     "@babel/preset-typescript": "^7.17.12",
-    "@gmod/indexedfasta": "^2.0.2",
+    "@gmod/indexedfasta": "^2.1.0",
     "@types/buffer-crc32": "^0.2.2",
     "@types/jest": "^29.2.4",
     "@types/long": "^4.0.2",
