+++ src/rans/d04.ts
@@ -1,3 +1,4 @@
+//@ts-nocheck
 import { CramMalformedError } from '../errors'
 
 import { TF_SHIFT } from './constants'
