+++ src/rans/decoding.ts
@@ -1,3 +1,4 @@
+//@ts-nocheck
 import { CramMalformedError } from '../errors'
 
 import { RANS_BYTE_L } from './constants'
