+++ src/rans/index.ts
@@ -1,3 +1,4 @@
+//@ts-nocheck
 import { CramMalformedError } from '../errors'
 
 import Decoding from './decoding'
