+++ test/compressions.test.ts
@@ -1,0 +1,31 @@
+//@ts-nocheck
+import { testDataFile } from './lib/util'
+import { dumpWholeFile } from './lib/dumpFile'
+import { CramFile } from '../src/index'
+import { FetchableSmallFasta } from './lib/fasta'
+
+test('lzma', async () => {
+  const fasta = new FetchableSmallFasta(testDataFile('hts-specs/ce.fa'))
+  const seqFetch = fasta.fetch.bind(fasta)
+  const file = new CramFile({
+    filehandle: testDataFile('hts-specs/0903_comp_lzma.cram'),
+    seqFetch,
+  })
+  const fileData = await dumpWholeFile(file)
+  const feat = fileData[2].data[1].features[0]
+  const hardClip = feat.readFeatures[0]
+  expect(hardClip).toMatchSnapshot()
+})
+
+test('bzip2', async () => {
+  const fasta = new FetchableSmallFasta(testDataFile('hts-specs/ce.fa'))
+  const seqFetch = fasta.fetch.bind(fasta)
+  const file = new CramFile({
+    filehandle: testDataFile('hts-specs/0902_comp_bz2.cram'),
+    seqFetch,
+  })
+  const fileData = await dumpWholeFile(file)
+  const feat = fileData[2].data[1].features[0]
+  const hardClip = feat.readFeatures[0]
+  expect(hardClip).toMatchSnapshot()
+})
