+++ test/dump.test.ts
@@ -79,42 +79,6 @@ test('works with hard clipping', async () => {
   expect(hardClip.data).toEqual(803)
 })
 
-test('lzma', async () => {
-  const fasta = new FetchableSmallFasta(testDataFile('hts-specs/ce.fa'))
-  const seqFetch = fasta.fetch.bind(fasta)
-  const file = new CramFile({
-    filehandle: testDataFile('hts-specs/0903_comp_lzma.cram'),
-    seqFetch,
-  })
-  const fileData = await dumpWholeFile(file)
-  const feat = fileData[2].data[1].features[0]
-  const hardClip = feat.readFeatures[0]
-  const nextReadFeature = feat.readFeatures[0]
-  expect(hardClip.refPos).toEqual(1050)
-  expect(nextReadFeature.refPos).toEqual(1050)
-  expect(hardClip.refPos).toEqual(1050)
-  expect(hardClip.pos).toEqual(51)
-  expect(hardClip.data).toEqual(1)
-})
-
-test('bzip2', async () => {
-  const fasta = new FetchableSmallFasta(testDataFile('hts-specs/ce.fa'))
-  const seqFetch = fasta.fetch.bind(fasta)
-  const file = new CramFile({
-    filehandle: testDataFile('hts-specs/0902_comp_bz2.cram'),
-    seqFetch,
-  })
-  const fileData = await dumpWholeFile(file)
-  const feat = fileData[2].data[1].features[0]
-  const hardClip = feat.readFeatures[0]
-  const nextReadFeature = feat.readFeatures[0]
-  expect(hardClip.refPos).toEqual(737)
-  expect(nextReadFeature.refPos).toEqual(737)
-  expect(hardClip.refPos).toEqual(feat.alignmentStart)
-  expect(hardClip.pos).toEqual(1)
-  expect(hardClip.data).toEqual(803)
-})
-
 function isIterable(input) {
   if (input === null || input === undefined) {
     return false
