+++ test/lib/dumpFile.ts
@@ -1,3 +1,4 @@
+//@ts-nocheck
 async function dumpSlice(container, sliceOffset) {
   const slice = container.getSlice(sliceOffset)
   const header = await slice.getHeader()
