+++ test/lib/syncLocalFile.ts
@@ -1,3 +1,4 @@
+//@ts-nocheck
 // don't load fs native module if running in webpacked code
 const fs = typeof __webpack_require__ !== 'function' ? require('fs') : null // eslint-disable-line camelcase
